backend/pyproject.toml
[tool.poetry]
name = "swallowtail-backend"
version = "0.1.0"
description = "Autonomous multi-agent e-commerce platform backend"
authors = ["Swallowtail Team"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
crewai = "^0.148.0"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.16"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic = "^2.7.1"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.0"
redis = "^5.0.4"
celery = "^5.3.4"
kombu = "^5.3.4"
httpx = "^0.27.0"
websockets = "^12.0"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
black = "^24.4.2"
ruff = "^0.5.1"
setuptools = "^80.9.0"
duckduckgo-search = "^8.1.1"
pinecone-client = "^5.0.1"
sqlalchemy = "^2.0.35"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
greenlet = "^3.1.1"
pillow = "^10.4.0"
supabase = "^2.9.2"
nest-asyncio = "^1.6.0"
python-socketio = {extras = ["asyncio"], version = "^5.13.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"
pytest-timeout = "^2.3.1"
black = "^24.4.2"
ruff = "^0.5.1"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "PT", "RET", "SIM"]
ignore = ["E501", "S101"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true